* main:
  Allow `drop_table` to accept an array of table names.
  Use json format for manifest URL (#52948)
  Address `DebugExceptionsTest` failures against Ruby with Prism parser
  Fix log message inconsistency in hash_password method
  Revert `params.fetch` to `params.expect` conversions in scaffold
  Prevent excess spacing
  Explain the manifest link better and hide by default
  Prevent excess CR
  Use 512x512 to match the manifest
  Remove sucker_punch ActiveJob::Queue adapter (#52935)
  [ci-skip][docs] Use 'change_column_null' instead of 'change_column_default'
  Do not include redis by default in dev container
  Add missing period in Active Storage guide.
  Include options when instrumenting ActiveSupport::Cache::Store delete
  Fix Rubocop errors for Rails apps without jbuilder
  Do not create `scan_js` workflow for API-only apps #52900
  Wrap database seeding in a reload
  bump solid_cable to fix tests
  Use thread pool for ActionController::Live
  Refactor implementation of attributes_for_inspect default values

+++ activerecord/lib/active_record/migration/command_recorder.rb
@@ -202,13 +202,18 @@ module ActiveRecord
         end
 
         def invert_drop_table(args, &block)
-          if args.last.is_a?(Hash)
-            args.last.delete(:if_exists)
+          options = args.extract_options!
+          options.delete(:if_exists)
+
+          if args.size > 1
+            raise ActiveRecord::IrreversibleMigration, "To avoid mistakes, drop_table is only reversible if given a single table name."
           end
-          if args.size == 1 && block == nil
+
+          if args.size == 1 && options == {} && block == nil
             raise ActiveRecord::IrreversibleMigration, "To avoid mistakes, drop_table is only reversible if given options or a block (can be empty)."
           end
-          super
+
+          super(args.push(options), &block)
         end
 
         def invert_rename_table(args)
